Move most of the Ganesh-specific logic out of SkSpecialImage.cpp
This introduces a namespace called SkSpecialImages which
allows us to define Ganesh- and Graphite-specific factories
in different files and remove a bunch of #ifdefs from SkSpecialImage.cpp

Because FilterResult::MakeFromImage needed to call a potentially
Ganesh-specific version of SkSpecialImages::MakeFromImage, I
introduced a delegate for making images (like we had for making
SkSpecialSurfaces).

The Graphite version of this delegate just produces a raster-backed
SkSpecialImage to preserve the same functionality as before.

Change-Id: I4a738f78c0e2c95539e66bf3a15c6b786e0393bd
Bug: skia:14317
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/722476
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>