[graphite] Fix bounds used to order clip draws
In https://skia-review.googlesource.com/c/skia/+/712637, we accidentally
switched to passing the 'outerBounds' parameter into the BoundsManager's
getMostRecentDraw() instead of 'fOuterBounds' of the Clip element.

The 'outerBounds' parameter is renamed to 'drawBounds' to better
highlight their differences, and it fixes the getMostRecentDraw() call
to use the clip's bounds for the query. This ensures that the clip
order determined for an element is valid regardless of any future draws
that change its used bounds. It fixes the ordering issues identified
in GM_strokerects when subjected to a rotated transform.

Change-Id: I766f8abf089ee529587fade0347c6cc388f85384
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/723079
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Arman Uguray <armansito@google.com>