[canvaskit] Add error check when loading SKP
When loading a single SKP the load status was not checked. This
resulted in a null ptr being added to the frames list and a
memory error later on.

This was encountered when running in a local debug environment
where the max picture version was less than the one being
served by fiddle.skia.org.

Bug: skia:none
Change-Id: Iaf27f4d15e93541c88989730a59615bc25efbddb
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/725556
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Chris Mumford <cmumford@google.com>