[skif] Implement image delegate for graphite
This also cleans up the not-so-necessary debug-only RectFits function.

This updates Ganesh's MakeFromTextureImage() to require a context, but preserves the policy of automatically uploading a raster image into a ganesh special image.

The new Graphite factory sort of follows this behavior, but relies on the client ImageProvider to be the final arbiter of whether or not the raster image is converted into a graphite image.

Bug: b/263138161
Change-Id: Ib03a081bf86b60aceb1da87746e30c16276ffa8d
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/725196
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>