[graphite] Fill out semaphore support for Vulkan swapchain.
To support the Vulkan swapchain we need to a wait and a signal
semaphore when inserting the Recording. The wait is used to stall the
rendering on the GPU until the acquired backbuffer surface is ready.
The signal is used to stall the swapchain present until the rendering
is completed.

Because the other backends don't need this setup, the Recording snap
has been moved out of WindowContext and into the individual Graphite
WindowContexts.

Bug: b/239826369
Change-Id: I7e5655c11950189c9797e039020de8858a64b88d
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/720913
Reviewed-by: Nicolette Prevost <nicolettep@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Jim Van Verth <jvanverth@google.com>