Fix data going empty for fuzz
The span named data, reached size 0, then front() was called.
This caused the newly added bounds checking call to trigger.

Add code to exit the fuzzer loop when data has no bytes, and
add a guard so that memcpy and front() are not called when there
are no bytes.

Bug: oss-fuzz:60689
Change-Id: I94f14cd194c66bec13e8dc10ca87aa20bdc38575
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/725717
Commit-Queue: Herb Derby <herb@google.com>
Reviewed-by: John Stiles <johnstiles@google.com>