[bazel] //gm/BazelGMRunner.cpp: Add support for the "picture_serialization" via.
The "picture_serialization" via is equivalent to DM's "serialize" via. The new name was chosen because it is more descriptive and because it makes its relation to the "picture" via more clear (see follow-up CL: https://skia-review.googlesource.com/c/skia/+/725878).

This CL extracts the process of drawing a GM and grabbing a bitmap out of the main() function and into the draw() function declared in //gm/vias/Draw.h. It also adds a new "via" argument to that function. Within that directory, multiple implementations for draw() exist in separate .cpp files. The idea is that each draw() implementation will only support a subset of vias that are related to each other (e.g. //gm/vias/SimpleVias.cpp will support the "direct", "picture" and "picture_serialization" vias). The motivation is that this will make the build more modular, as related vias usually share the same set of dependencies (e.g. SkPictureRecorder in the case of the "picture" and "picture_serialization" vias).

Tested with the following Bazel invocations:

    $ bazel test //gm:cpu_8888_via_picture_serialization_test \
        --config=linux_rbe \
        --strategy=TestRunner=local

    $ ssh -L 5037:localhost:5037 skia-rpi2-rack1-shelf1-036
    $ bazel test //gm:gpu_gles_via_picture_serialization_android_test \
        --config=linux_rbe \
        --config=pixel_5 \
        --config=gl_ganesh

Bug: skia:14227
Change-Id: Ifa272a76226f80b99da90a53dcce6edaa4aed12b
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/725857
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Leandro Lovisolo <lovisolo@google.com>