Fix WGSL codegen for intrinsics that might overflow.
This requires different approaches depending on the intrinsic.
For intrinsics with a native WGSL equivalent, we interpret calls
where all arguments are compile-time-constant as an overflow
risk. (Normally, an intrinsic with all compile-time-constant
inputs would have been evaluated at compile time, but we would
have left it as-is if the result overflows.) We can defuse the
WGSL compile-time checks by replacing any argument with a let-
variable; this turns the expression into runtime-evaluated
instead of compile-time evaluated. In practice, we replace the
first argument with a _skTemp value.

For intrinsics where we polyfill the intrinsic, the fix needs
to be handled on a case-by-case basis depending on the contents
of the polyfill. e.g. if the polyfill has `arg0 * arg1`, we
need to make sure that we replace either arg0 or arg1 with a
scratch-let. This is fiddly work, but I've added tests for
each, and fortunately there aren't many cases like this.

Bug: skia:14385
Change-Id: Id882f2bb2ab6e165bea6ce654d24986a85b4c8e6
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/727258
Commit-Queue: Arman Uguray <armansito@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Reviewed-by: Arman Uguray <armansito@google.com>