Eliminate fInterfaceBlockMap from Metal codegen.
This map was only used to associate a field with an interface block
when emitting field-access expressions. This allowed us to give a
unique name to anonymous interface-blocks when their field accesses
occurred. Instead of having a map, we can use the type of the
interface block as a key for this lookup instead, since each
interface block must have a unique type.

Change-Id: Ief0beed539ff215c0a8f2fe39f62b86983a2de55
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/728916
Reviewed-by: Brian Osman <brianosman@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>