Avoid sk_sp to static local variable
SkEmptyFontMgr is the last resort fallback for font managers. Prior to
this change if one was needed it would be created as a static local
variable and an sk_sp to it returned and then the sk_sp was also stored
as a static local variable. This meant that the SkEmptyFontMgr static
local could be destroyed before the sk_sp static local variable. This is
the Static Initialization Order Fiasco but for destruction.

To avoid these problems create the single SkEmptyFontMgr on the heap if
it is needed and make it always owned by sk_sp. The order of destruction
is no longer a problem as the last sk_sp to be destroyed will delete the
SkEmptyFontMgr.

Change-Id: Ia32678ec449dbfc4674a276193b636c29dbaf9c2
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/728738
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>
Auto-Submit: Ben Wagner <bungeman@google.com>
Commit-Queue: Leandro Lovisolo <lovisolo@google.com>