Roll back Nvidia drivers on Skolo Linux machines with RTX 3060.
In https://skia-review.googlesource.com/c/skia/+/729801 I tried to update the Nvidia drivers from 470.141.03 to 535.86.05. We abandoned that upgrade for now because the new driver apparently introduces a memory leak.

When I tried to roll back to the original drivers, which are installed from the Debian Bullseye repositories, it reverted to version 470.182.03, which is not the same as the version we started from (470.141.03). This CL updates the gpu dimension of the affected tasks accordingly.

Bug: b/40045524
Change-Id: I5323ddf8867bd9f9b092ab4d5ea281a51492d9aa
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/729803
Reviewed-by: John Stiles <johnstiles@google.com>
Commit-Queue: Leandro Lovisolo <lovisolo@google.com>