Prime expat's hash salt on Windows
When jpegs have gainmaps expat is used to parse out the information.

As part of this, expat initializes its hash seed by calling rand_s.

On Windows this results in a call to cryptbase!RtlGenRandom which we
would like to avoid loading in sandboxed renderer processes.

This CL directly primes expat using XML_set_hash_salt() with an
address-derived value - this should be sufficient for the DOS
mitigation that the salt provides.

Bug: 74242,1466200
Change-Id: Ide0fe82fb1a3158f07d06d7e35093b16e4e363fd
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/730076
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Alex Gough <ajgo@chromium.org>