Avoid mutating parameter flags after creation.
Previously, when creating a function declaration, we would iterate
over the parameters and convert a parameter like `in float x` to
`float x`, because the `in` flag on a parameter is the default
option and this makes it easier to detect duplicate definitions.

We now perform this flag tweak when the variable is created; we
already know it's a parameter (via its Storage type) and so we
don't actually need to mutate it at all.

(This allows us to remove `setModifiers` from Variable; keeping
things immutable where we can is a nice property.)

Change-Id: I52aa9dc1ecb34100147ca7542e95f2f7e4243578
Bug: b/40045537
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/731485
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>