Split Variable modifiers() into layout() and modifierFlags().
The trick to removing the ModifiersPool is that most variables don't
use a layout. We already have divided Variables into two classes; the
smaller `Variable` and the full-sized `ExtendedVariable`. Only
`ExtendedVariable` needs to store a layout; the smaller `Variable`
class can just assume a default-initialized layout with no fields
set.

One complication here is that the Variable API didn't expose the
layout as its own accessor; it exposed `modifiers()`, which is a
combination of the modifier flags (which vary often) and the
layout (which is rarely used). This meant that a call to the
`modifiers()` accessor in the common case would need to return
something that we didn't have on hand, and would have needed to
synthesize a Modifiers struct containing both the variable's
actual ModifierFlags and an empty Layout.

To solve this complication, this CL removes the
`Variable::modifiers` accessor, and replaces it with two accessors,
`Variable::modifierFlags` and `Variable::layout`. In practice this
actually makes a lot of the code read better anyway.

Change-Id: I2aab9898d6905d078752f9328d85d9db6c1eb070
Bug: b/40045537
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/731599
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>