[skif] Refine ShaderFlags and Builder::eval API
ShaderFlags::kExplicitOutputBounds is gone, passing in a bounds value
to eval() is sufficient.

ShaderFlags::kOutputIsUnion is gone, it was only used in merge() so
that just calculates the union directly. Also added a Union template
and updated a few image filter bounds functions to use it.

Renamed kNonLinearSampling to kNonTrivialSampling to imply less
about how the sampling is structured.

Removed ShaderFlags::kSampleInParameterSpace and removed the
flags passed into eval(). It turns out the only flag that had to be
passed in to eval() to control the effect shader was this one. All
the others were really just to control the input shader behavior, so
add() is the only way to do so. And the behavior required for
kSampleInParameterSpace wasn't really a shader flag because it also
impacted the coordinate space of the canvas being drawn into. Now the
logic for that has been removed from FilterResult::asShader() and
is handled entirely in the Builder, but should be equivalent.

Change-Id: I45d31d7b80b83d574f41564aded7ce4926487ad8
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/731816
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>