[OpenGL] Another approach to fixing mipmap generation on Intel.
Setting BASE_LEVEL equal to MAX_LEVEL seems to be a signal to Intel devices to only sample within that single miplevel, which avoids some
crosstalk with the next level up that's being rendered to. However,
doing that violates the GL spec wrt setting up the framebuffer at MAX_LEVEL+1, and hence fails in the ANGLE validator. The only solution
in that case seems to be to set up syncs between the render of each
level.

Bug: chromium:1463432
Bug: b/40045272
Change-Id: Id5560984356b1d99e2c773fa719cea24e3a2bfff
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/733076
Commit-Queue: Jim Van Verth <jvanverth@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>