Remove SK_GL #ifdefs from GrBackendSurface
This outlines the removal of the union fields in GrBackendSurface
and replaying them with a pointer to a subclassable GrBackend*Data
type. This allows us to remove all SK_GL #ifdefs from the
GrBackendSurface code, allowing us to decouple each of the the
Ganesh backends from one another.

In order to be able to downcast GrBackendFormatData, it needs to be a
pointer, which means the GrBackendFormat et al are no longer purely
on the stack, but have a unique_ptr to the data of the correct subclass.

Change-Id: Ib5f305a82ddd5cb995a3174b008812780eef21fc
Bug: b/293490566
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/701398
Commit-Queue: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Greg Daniel <egdaniel@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>