Move GrAHardwareBufferUtils to include/android and split GL/VK parts
Right now, Android #include src/gpu/GrAHardwareBufferUtils.h, which
is not compatible with the idea of a modular build and strictly
enforced include visibility. This follows the approach our other
Android-only files use, moving to include/android.

I also added some explicit GL and Vulkan functions so we can
remove uses of #ifdef SK_GL etc in the Skia codebase. We will
need to update Android to call the correct variant (or bail if
it was not compiled with the GL or Vulkan backend).

Change-Id: I6b297b0b6f0b209f097307c3120e2dbd3609ff13
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/735976
Commit-Queue: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>