Add an EmptyExpression IR element to SkSL.
When the inliner replaces a void-typed function call, it needs to
replace the expression with _something_. Since there's no such
thing as a void-typed value in GLSL, we didn't have an existing
IR node that we could use here, so SkSL arbitrarily inserted a
`false` node since it seemed safe. However, in this test case, there
is an outer void-typed comma-expression; it does not get a chance
to re-evaluate its type after the inliner replaces its void-typed
RHS value with a bool-typed value, and still considers itself to
be void-typed, despite evaluating to a value (`false`).

RP would get confused when emitting expressions because the outer
comma-expression's type (`void`: zero slots) mismatched its
actual pushed slot-count (`false`: one slot). RP would push
the `false` (+1 slot) but pop nothing because of the void-typed
outer expression,  unbalancing the stack.

We now have a dedicated EmptyExpression which we can use as a
replacement expression for the inliner. It has zero slots.

Bug: oss-fuzz:61195
Change-Id: Idd238e52ad644911ea58d8d43d6dffcb8f365b68
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/736716
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>