Reject sequence-expressions containing arrays.
Our existing strict-ES2 check here was insufficient, because it
assumed the LHS and RHS of the binary expression were of compatible
types, and so we checked only the left hand side. The assumption of
compatible types is usually true for most binary expressions, but it
is not true for a sequence-expression.

WebGL2 also disallows sequence-expressions with arrays. There's
very little to gain by allowing them, and lots of GLSL
implementations struggle with sequence-expressions in general,
so SkSL now unilaterally rejects sequence-expressions with arrays,
regardless of the strict-ES2 setting.

Bug: b/294893925
Change-Id: I2a1d765dbf34724a0d0a4428e44d6217036c7911
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/737816
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>