Propagate protected-ness to dependant draws
For example, when blurring a protected image we need to create protected surfaces and SFCs to accommodate the protectedness.

Additionally, having an API-level entry point to create a protected SkSurface seems like it would be useful to any client trying to work with protected images or trying to create protected content.

Change-Id: I191f50ec1e04175c3e243de177ba21e0970f2a58
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/738160
Reviewed-by: Brian Osman <brianosman@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>