Add workaround for WebGL2 and void-typed sequence expressions.
We now have a shader caps bit for `fCanUseVoidInSequenceExpressions`.
Out of an abundance of caution, this bit is always set to false in
WebGL. When this is false, functions with a void return type are
emitted with a float return type instead, and `return` statements
are rewritten as `return 0.0;`. This should be relatively harmless,
and makes these function calls eligible to participate in sequence
expressions (since they are no longer void-typed).

No intrinsics in sksl_shared return void. The only private intrinsic
I noticed with a void return type was `atomicStore` (which shouldn't
be available to WebGL).

Bug: b/294893925
Change-Id: Ibff853b46f2134e29ace5274f344541aa40d087a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/737818
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>