Write the WGSL entrypoint function after writing all ProgramElements.
Previously, we would write the entrypoint as soon as main() was
encountered. This is generally last, but that's not guaranteed; you
could easily use prototypes to structure your code so that main() is
near the top.

This makes it safe to assume in `writeEntryPoint` that the entire
program has been processed.

Change-Id: If4c09d20c967dd7866b659ee184fdc532c714aec
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/739858
Commit-Queue: James Godfrey-Kittle <jamesgk@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Reviewed-by: James Godfrey-Kittle <jamesgk@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>