[bazel] Add the "zip" command to the Linux RBE toolchain container image.
The "zip" command needs to be present in the RBE toolchain container in order for Bazel to be able to zip up any undeclared test outputs. This is the mechanism we use to retrieve PNGs generated by GM tests that run on RBE (e.g. those that do not require a GPU).

Bug: b/40045301
Change-Id: I8cd9de6fead5380dacd4b899e1da9cf91ab2990a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/740397
Commit-Queue: Leandro Lovisolo <lovisolo@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>