[ganesh] Fix asyncReadPixels when running on Direct3D
When performing pixel conversions, we were passing in the dstInfo
rowBytes to GrConvertPixels, but then using the padded transfer buffer
rowBytes to read from the converted data. Changed to explicitly
specify what the rowBytes of the final transfer data will be.

This is the same bug and fix as https://review.skia.org/671057.

Change-Id: Ib75f7b8d0a1b227b213db568c6fefa439840ab0a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/741818
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Jim Van Verth <jvanverth@google.com>