[bazel] //gm/BazelGMRunner.cpp: Simplify how Gold keys are generated.
This CL moves the "name" key to GM::getGoldKeys(), and the "surface_config" key to SurfaceManager::getGoldKeys(). The goal is to have all related keys be determined in a single location, which makes logical sense to me and will hopefully make things easier to follow for others.

Bug: b/40045301
Change-Id: I9cb3973ebd886ed7ecdbd3344f5d59a07acb618f
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/741817
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Leandro Lovisolo <lovisolo@google.com>