Perform error-checking when creating a Dawn shader module.
Previously, we would check if `CreateShaderModule` returned a null
wgsl::ShaderModule, but this is not how errors are reported in
Dawn. Instead, we need to call `module->GetCompilationInfo` and
look for any `WGPUCompilationMessageType_Error`-typed messages.
If we find any, we now invoke the ShaderErrorHandler and stop
the pipeline creation.

Previously, attempting to use a malformed shader would lead to a
cascade of errors in Debug and a crash in Release.

Change-Id: If7b221e9b293f8d67007c6caf182d0f5ebadebcd
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/742739
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Reviewed-by: James Godfrey-Kittle <jamesgk@google.com>