Convert skqp_app to skqp_jni
These are essentially the same thing - a JNI library used by either CTS
or an APK. In fact, the CTS build target is constructed from one and
then gets the other name. For clarity, it makes sense to use a single
name.

In addition, my local attempts to build an APK resulted in an app that
crashed as soon as we tried to run the tests, because the app was
trying to load libskqp_jni, but the thing we built was libskqp_app. I'm
not sure how this worked before, but it may not have for a while. Using
the same name makes this work.

Splitting out from I62ddd379912e67b62db382f72da485c672f8edfc as this is
an isolated change that does more than getting SkQP building locally.

Bug: 285581014
Change-Id: I53eabdd55eb0ad9746d5c6e9fdb0ab92cb6a8a3a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/743476
Reviewed-by: Derek Sollenberger <djsollen@google.com>
Commit-Queue: Leon Scroggins <scroggo@google.com>