GraphiteDawn: use Dawn's MSAARenderToSingleSampled feature.
When available, this feature will implicitly render to a single sampled
texture with multisampling. There is no need to manually allocate any
MSAA textures when using this feature.
This feature also helps avoid an extra copy when loading the resolve
texture. The extra copy path will still be the fallback if the feature
is not available.

Bug: dawn:1710
Change-Id: I50b1047c6d9dff20d58417c34ea88e80b88edaf2
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/709137
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Quyen Le <lehoangquyen@chromium.org>