[graphite][compute] Dynamically assemble resource declarations from snippets
To correctly support the nuances of binding index assignment across the
various API backends, ComputeSteps now specify only the variable
declarations in SkSL snippets attached to each resource declaration.
Type and layout declarations are handled by context utilities when the
shader gets assembled during pipeline creation.

This brings ComputeSteps closer to RenderSteps when it comes to resource
declarations, at the cost of making the program definition a little less
centralized.

There is currently no affordance for type definitions that may be
referenced by a resource declaration (e.g. a SSBO that contains an
array of a custom struct type doesn't have a way to declare a custom
struct). This may be supported in the future using an abstraction like
ComputeStep::headerSkSL() to place common type definitions before
resource declarations.

The Compute_SampledTexture test is now enabled on Dawn contexts as
buffer, texture, and sampler indices are now all assigned correctly.
This brings compute support on the Dawn backend to parity with the Metal
backend.

Bug: b/262427430
Change-Id: I285c366a44c7e89c96649d3c379a7abf939318b1
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/743637
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Arman Uguray <armansito@google.com>