Remove GrBackendRenderTargetData::isValid
This was only being checked when copying, and didn't really match the
behavior before the removal of SK_GL ifdefs. If the old behavior is
important (ensuring isValid is false when constructed from a bad
format), we should figure out how to do that at initialization time.

Bug: b/293490566
Change-Id: I19b12a272dd6a9be63082ca07cbf4dc7f752bde3
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/744661
Commit-Queue: Brian Osman <brianosman@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>