Move 1D linear Gaussian FP to a runtime GrSkSLFP
Like in https://skia-review.googlesource.com/c/skia/+/746856, the
Gaussian blur is now a runtime effect. It's a single effect with a
constant loop, making it equivalent to what the FP produced when in
reduced shader mode. If there are performance regressions we can bring
back some amount of binning based on the radius.

GrBlurUtils now handles creating the GrTextureEffect that binds to the
child uniform and removes the logic for trying to strip the domain,
etc. This was likely the code path being taken anyways for special
images but if there are regressions, then all the same information is
available to create the domain-optimized texture effect in GrBlurUtils.

Bug: b/294102906
Change-Id: If3d7f748e921253a4e7912fb2a56cbfa408a1dcc
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/746857
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>