Avoid generating unnecessary _skParam variables in WGSL.
These only added value in the case where the parameter was
modified, since WGSL parameters are immutable (we would make a
var-based copy from `_skParamN` to the expected var name).

Previously we would always create the _skParam name and use `let`
to alias the parameter to its actual name. Theoretically, in WGSL
a let-assignment is "free"--but in practice we see the final Metal
output still contains these extra copies, so it seems worthwhile
to try and eliminate them when we can.

It's relatively easy for us to avoid emitting them, and relatively
much more difficult for the final shader compiler to prove that
it's safe to optimize them away.

Change-Id: I757738c0416f0ee05e42a4e2b28347cb4c64b84d
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/748319
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>