Move SkOpts things around for easier modularization
It's a bit awkward to have a src/core depend on src/codec,
so the swizzles that are needed in core are now defined in
src/SkSwizzlePriv.h.

Additionally, it's a bit awkward to have many of the SkOpts
implementations actually depend on SkOpts due to the indirect
#include of SkRasterPipelineOpList.h, so I split out the
defines that are necessary into their own file.

Change-Id: Iba4efdd756996b234fb6dfa593e0e1ff444d6f20
Bug: b/284138320, b/40045064
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/748876
Reviewed-by: Brian Osman <brianosman@google.com>