Exact loops for all 1D blur kernels
This is a speculative CL, to land, collect perf numbers, and then
rollback so we don't have to worry about staging into clients. There
are some simplifications to the SkSL that might be nice to preserve
if possible, but I would prefer not to have 28 pipelines per blur
even if it turns out that the loop branching was the cause of the
regression (which this CL entirely removes).

Assuming it doesn't remove the regression, then we will try declaring
tighter uniform arrays.

Bug: b/297590025
Change-Id: I67aa73148777fe9be2a79ccf16cbb9f70568ae4e
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/749059
Reviewed-by: Robert Phillips <robertphillips@google.com>
Auto-Submit: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>