[skif] Fix small-sigma threshold on GPU blurs
https://skia-review.googlesource.com/c/skia/+/746817 had a copy-pasta
error when I moved the maxOutput calculation back to before the split
between GPU and CPU blurs; it was passing gpuBacked=false to the bounds
functions instead of ctx.gpuBacked(). That was okay when it was only
meant for the temporary CPU impl, but not switching to GPU-backed
meant that small sigmas that the GPU knew how to blur were being
identified as "identity" and then clipped.

Bug: b/294102906
Change-Id: I99c0e6f29e60162e35c92a1e74d2cae500875125
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/749736
Reviewed-by: Brian Osman <brianosman@google.com>
Auto-Submit: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>