Disallow `texture2D` without readonly/writeonly qualifier.
This was a small code change but had a substantial blast radius
on test code; many generic test programs included unqualified
texture2Ds.

This was also the first time that an empty layout/modifier set
could be flagged as invalid code. Flagging a zero-length range
of code led to poor diagnostics (the ^^^^ range where modifiers
should be highlighted was missing; since there were none).
In this case, we now highlight the type itself as the error
location instead of the modifier range.

http://go/samplers-and-storage-textures-in-sksl#heading=h.ia142csj1140

Bug: b/295161016
Change-Id: I3b16780efc9860a751b3e45f64fd4afb2219998a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/750358
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Arman Uguray <armansito@google.com>